fix: fixed announce highscore
before, the players highscore was updated before it was announced. The highscore will only be announced when the currentpointf of the players are higher than the players highscore. The highscore was already updated to the currenpoints and so the highscore is equal to the currentpoints in the announcement check causing the highscore to not be announced to the player